+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -332,6 +332,7 @@ where
                 storage_path,
                 TxStatusBlobSenderHooks::new(tx_status_manager.clone()),
                 shutdown_sender,
+                Duration::from_secs(config.blob_processing_timeout_secs),
             )
             .await?;