+++ crates/module-system/sov-test-utils/src/lib.rs
@@ -101,6 +101,9 @@ pub type TestStorageManager =
 /// The maximum batch size that the preferred sequencer can create.
 pub const TEST_MAX_BATCH_SIZE: usize = 1024 * 1024;
 
+/// I a blob is not processed within this time, it will shut down the sequencer.
+pub const TEST_BLOB_PROCESSING_TIMEOUT: u64 = 60;
+
 /// The maximum number of concurrent blobs.
 pub const TEST_MAX_CONCURRENT_BLOBS: usize = 16;